Need 64 bit ints

+++ 2021/c/day24/day24.c
@@ -5,7 +5,7 @@
 #include <signal.h>
 #include <err.h>
 
-int monad(int input[14]);
+int64_t monad(int input[14]);
 
 static void
 solve(int job)
